Refactor implementation: split into several files
