+++ tool/release/BUILD
@@ -30,16 +30,6 @@ py_binary(
 )
 
 kt_jvm_binary(
-    name = "create-notes",
-    srcs = ["CreateNotes.kt"],
-    main_class = "com.vaticle.dependencies.tool.release.CreateNotesKt",
-    deps = [
-        "@maven//:com_eclipsesource_minimal_json_minimal_json",
-        "@maven//:com_google_http_client_google_http_client",
-    ],
-)
-
-kt_jvm_binary(
     name = "bump",
     srcs = ["Bump.kt"],
     main_class = "com.vaticle.dependencies.tool.release.BumpKt",
